Split off most encoding-related tests to a separate file
Write a few new ones

Fix up descriptions on some decoding functions
