+++ tests/documents/utf8.xml
@@ -1,0 +1,3 @@
+﻿<?xml version="1.0"?>
+<project name="project-name">
+</project>
